try api as context manager

+++ tests/test_multi_namespace_spawner.py
@@ -7,12 +7,13 @@ import pytest
 from jupyterhub.objects import Hub
 from jupyterhub.objects import Server
 from jupyterhub.orm import Spawner
-from kubernetes.client import V1Namespace
-from kubernetes.client.models import V1Capabilities
-from kubernetes.client.models import V1Container
-from kubernetes.client.models import V1Pod
-from kubernetes.client.models import V1SecurityContext
-from kubernetes.config import load_kube_config
+from kubernetes_asyncio.client import V1Namespace
+from kubernetes_asyncio.client.models import V1Capabilities
+from kubernetes_asyncio.client.models import V1Container
+from kubernetes_asyncio.client.models import V1Pod
+from kubernetes_asyncio.client.models import V1SecurityContext
+from kubernetes_asyncio.client.rest import ApiException
+from kubernetes_asyncio.config import load_kube_config
 from traitlets.config import Config
 
 from kubespawner import KubeSpawner
@@ -62,20 +63,20 @@ async def test_multi_namespace_spawn():
 
     # get a client
     kube_ns = spawner.namespace
-    load_kube_config()
-    client = shared_client('CoreV1Api')
+    await load_kube_config()
+    client = await shared_client('CoreV1Api')
 
     # the spawner will create the namespace on its own.
 
-    # Wrap in a try block so we clean up the namespace.
+    # Wrap in a try block so we clean up the namespace in finally.
 
-    saved_exception = None
     try:
         # start the spawner
         await spawner.start()
 
         # verify the pod exists
-        pods = client.list_namespaced_pod(kube_ns).items
+        p_list = await client.list_namespaced_pod(kube_ns) 
+        pods = p_list.items
         pod_names = [p.metadata.name for p in pods]
         assert "jupyter-%s" % spawner.user.name in pod_names
         # verify poll while running
@@ -84,15 +85,21 @@ async def test_multi_namespace_spawn():
         # stop the pod
         await spawner.stop()
         # verify pod is gone
-        pods = client.list_namespaced_pod(kube_ns).items
+        p_list = await client.list_namespaced_pod(kube_ns) 
+        pods = p_list.items
         pod_names = [p.metadata.name for p in pods]
         assert "jupyter-%s" % spawner.user.name not in pod_names
         # verify exit status
         status = await spawner.poll()
         assert isinstance(status, int)
-    except Exception as saved_exception:
-        pass  # We will raise after namespace removal
     # remove namespace
-    client.delete_namespace(kube_ns, body={})
-    if saved_exception is not None:
-        raise saved_exception
+    finally:
+        # Allow opting out of deletion.
+        if not os.environ.get("KUBESPAWNER_DEBUG_NAMESPACE"):
+            try:
+                await client.delete_namespace(kube_ns, body={})
+            except ApiException as exc:
+                if exc.status == 404:
+                    spawner.log.warning(f"Namespace {kube_ns} not found.")
+                else:
+                    raise
